Refresh volumes on terminating pods
https://github.com/kubernetes/kubernetes/issues/116481 describes an
issue where volumes stop being refreshed once a Pod enters the
"Terminating" state. This is a problem for serviceAccount tokens being
supplied via projected volumes. The net effect is that if a Pod is in
the terminating state for more than 20% of the expiration time of the
service account token, it may go stale. This is becoming a more
significant issue with the change to limit Kubernetes service account
tokens to 1 hour - since a terminationGracePeriod seconds of just 12
minutes could cause the token to become invalid. It's also a challenge
for other serviceAccount use cases, such as AWS IAM Roles - where the
max lifetime of the token of 1 day means that grace periods greater than
4.8 hours could result in an expired token.

My understanding of the way that volume refreshes work is that:

1. Once per second, syncLoop() will generate a tick
2. The tick causes the Kubelet's dispatchWork() method to be called
3. dispatchWork() in turn calls UpdatePod()
4. UpdatePod() makes the update available by assigning it to
   status.pendingUpdate and signals the availability of the update by
   writing to the podUpdates channel.
5. podWorkerLoop() handles the pod update
  a. If the Pod is running:
    * Calls SyncPod()
    * SyncPod() in turn calls the VolumeManagers's WaitForAttachAndMount().
      (but only if the container isn't Terminating - however, I don't
      think SyncPod is called if the container is Terminating so I don't
      understand this check)
    * WaitForAttachAndMount() will then call the
      DesiredStateOfWorldPopulator's ReprocessPod()
    * ReprocessPod() will mark Pod processing as failed
    * All this time, the VolumeManager has also been running a loop -
      once every 100ms it checks to see which volumes need to be
      reprocessed. Specifically its findAndAddNewPods() method will
      call processPodVolumes() - but only for Pods that aren't
      terminating.
    * processPodVolumes() will, if pod processing is not marked ats
      complete (which it isn't, because that's the net effect of
      ReprocessPod()) finally call the volume code that will cause
      projected volumes to be refreshed.
  b. If the Pod is Terminating:
    * SyncTerminatingPod() is called instead of SyncPod()
    * SyncTerminatingPod() calls killPod() that wait for the Pod to
      exit and sends signals as necessary
    * While waiting for the container to exit, it does no other work -
      specifically, it doesn't call WaitForAttachAndMount() like
      SyncPod() does.

So, looking at the problem reported in the issue, what I think is
happening is that while the container is running that chain of calls
results in the VolumeManager refreshing the volumes once per second.
However, once the Pod enters the Terminating state, we get stuck waiting
for the Pod to exit and we stop making the calls necessary for the
VolumeManager to refresh the volumes for us.

One way that we could handle this would be to call
WaitForAttachAndMount() once a second while killPod() is running. In
addition to this change, we also need to update the VolumeManager to not
ignore Terminating Pods. This approach works - in that volumes are
refreshed. However, this feels like an unfortunate solution: normally
WaitforAttachAndMount() is called in response in ticks from syncLoop() -
but with this change, we'd also be calling it during termination from a
completely different loop. Having two very different paths calling this
loop during different Pod lifecycles seems both confusing and error
prone.

So, instead, this commit takes a different approach:
* Instead of using ReprocessVolume() to signal to the VolumeManager when
  it's time to reprocess a Pod's volumes, we instead reprocess volumes on
  every loop of the VolumeManager.
* The above change means that we're going from reprocessing volumes once
  a second to 10 times a second. To account for this, we slow down the
  VolumeManager's loop to only run once a second as well.
* We remove the checks from the VolumeManger that prevent processing
  Terminating Pods.
* With these changes, we no longer need the ReprocessPod() or
  markPodProcessingFailed() methods - so we remove them.
* Tests are updated to reflect the method removals - and also to account
  for the fact that we now expect volumes to be reprocessed on
  Terminating Pods.

I think the nice thing about this approach is that handles the Running
and Terminating state uniformly - as long as the Pod is registered with
the VolumeManager, the volumes are updated. As I understand it, it won't
cause additional work because we also down the VolumeManager loop.

One thing that concerns me about this approach is that I do not fully
understand the conditions that I removed from findAndAddNewPods() in
order to make things work. These changes do have the intended effect of
making sure that Pods in the Terminating state continue to have their
volume's refreshed - but without fully understanding the conditions I
updated here I'm not 100% sure they don't have some other unintended
effect.