Refresh volumes on terminating pods
https://github.com/kubernetes/kubernetes/issues/116481 describes an
issue where volumes stop being refreshed once a Pod enters the
"Terminating" state. This is a problem for serviceAccount tokens being
supplied via projected volumes. The net effect is that if a Pod is in
the terminating state for more than 20% of the expiration time of the
service account token, it may go stale. This is becoming a more
significant issue with the change to limit Kubernetes service account
tokens to 1 hour - since a terminationGracePeriod seconds of just 12
minutes could cause the token to become invalid. It's also a challenge
for other serviceAccount use cases, such as AWS IAM Roles - where the
max lifetime of the token of 1 day means that grace periods greater than
4.8 hours could result in an expired token.

While a pod is running, SyncPod() will be called every now and then.
When it is called, it will then call WaitForAttachAndMount(). The result
of that call is that processPodVolumes(), when it runs, will call
MarkRemountRequired() for those pods. This results in volumes being
remounted. This is important for projected volumes since that is how
their contents are updated.

When a pod is in the terminating state, however, instead of SyncPod(),
SyncTerminatingPod() will run. And that function attempts to terminate
the pod and waits for that to complete - significantly without calling
WaitForAttachAndMount() at all - which means that projected volumes
won't be regenerated.

The approach that this commit takes to resolve that is that we update
findAndAddNewPods(). Previously, we'd ignore any terminating pods.
However, now if a pod is terminating, we call MarkRemountRequired().
Since we don't want to call that multiple times per second, we keep
track of the last time that we called that function and only call it if
its been at least 10 seconds since the last call.