Update the approach to address review feedback
The feedback was that manipulatring processedPods directly and calling
MarkRemountRequired() from within findAndAddNewPods() appeared to be
unsafe. Furthermore, it would be better to keep the code that does the
above functions co-located as opposed to having it split up into
multiple places.

So, in order to address that, we remove the above concerning code so
that all calls to MarkRoumountRequired() continue to be done in
processPodVolumes() as before.

The challenge with the above, is that:

1. processPodVolumes() is not called for terminating pods

2. In processPodVolumes(), we don't do anything if
   podPreviouslyProcessed() returns true. And podPreviouslyProcessed()
   will always return true since, for terminating pods, the pod worker
   loop stops calling ReprocessPod()

We address the first issue, by removing the check that skips calling
podPreviouslyProcessed() for terminating pods. I don't believe this will
have a performance impact since processPodVolumes() will immediatley
return if podPreviouslyProcessed() returns true - which it will unless
ReprocessPod() is called, as usual.

The second issue we address by modifying how podPreviouslyProcessed()
works. Previously, it would return true until markPodProcessingFailed()
was called (by way of ReprocessPod). We update the method so that now it
will return true only if its been less than a minute since
markPodProcessed() has been called. I don't believe that this introduces
any performance issues, since while a pod is running, the pod runner
loop calls ReprocessPod() a few times a minute anyway.

Anyway, the above resolves the issue where the pod loop stops calling
ReprocessPod() during pod termination which means that
podPreviouslyProcessed() will return true forever by changing it so that
every 1 minute, it will automatically start returning false, allowing
the remounting of volumes to occur.