Bail out of processPodVolumes() if the Pod is terminating
In a previous commit, we removed the check for if a Pod was terminating
from findAndAddNewPods(). Doing so is required, because that check
prevented us from calling processPodVolumes() - which is where we wanted
to mark volumes as needing to be refreshed, even when a Pod is
terminating. However, the issue is that with that check removed, we
might end up adding new volumes for not yet started containers in Pods
that are in the terminating state - which we don't want to do since
those containers will never actually start.

In order to fix the above problem, we move the check for if a Pod is
terminating into processPodVolumes() itself. If a Pod is terminating, we
will still make volumes as needing to be refreshed. However, we then
bail out of the method before adding any additional volumes - which I
belive resolves the issue.